refactor(RandomNumGenerator): util 패키지로 이동
