+++ src/main/java/lotto/util/RandomNumGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.util.Constant;
