+++ src/test/java/lotto/RandomNumGeneratorTest.java
@@ -3,7 +3,7 @@ package lotto;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.domain.RandomNumGenerator;
+import lotto.util.RandomNumGenerator;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
